Refactor and improve caching

+++ src/Platform/AngularDotnetPlatform.Platform/Caching/PlatformCacheEntryOptions.cs
@@ -14,12 +14,12 @@ namespace AngularDotnetPlatform.Platform.Caching
         /// Gets or sets how long a cache entry can be inactive (e.g. not accessed) before it will be removed.
         /// This will not extend the entry lifetime beyond the absolute expiration (if set).
         /// </summary>
-        public virtual int? SlidingExpirationInSeconds { get; set; }
+        public virtual double? SlidingExpirationInSeconds { get; set; }
 
         /// <summary>
         /// Gets or sets an expiration time after seconds, relative to now.
         /// </summary>
-        public virtual int? ExpirationInSeconds { get; set; } = DefaultExpirationInSeconds;
+        public virtual double? ExpirationInSeconds { get; set; } = DefaultExpirationInSeconds;
 
         /// <summary>
         /// Gets or sets an absolute expiration time, relative to now.
@@ -48,8 +48,8 @@ namespace AngularDotnetPlatform.Platform.Caching
             Configuration = configuration;
         }
 
-        public abstract override int? ExpirationInSeconds { get; }
+        public abstract override double? ExpirationInSeconds { get; }
 
-        public abstract override int? SlidingExpirationInSeconds { get; }
+        public abstract override double? SlidingExpirationInSeconds { get; }
     }
 }
